rest-api-to-postgresql/requirements.txt
appnope==0.1.3
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
debugpy==1.6.0
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
greenlet==1.1.2
idna==3.3
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.11
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pyzmq==23.1.0
requests==2.28.0
six==1.16.0
SQLAlchemy==1.4.37
stack-data==0.3.0
tornado==6.1
traitlets==5.3.0
urllib3==1.26.9
wcwidth==0.2.5
yarg==0.1.9
